server-side/requirements.txt
Flask
flask-login
argon2-cffi
